requirements.txt
ultralytics==8.2.103
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
seaborn==0.13.2
albumentations==1.4.15
streamlit==1.38.0
plotly==5.22.0
scikit-learn==1.5.1
PyYAML==6.0.2
tqdm==4.66.5
